+++ app/Http/Controllers/Auth/ForgotPasswordController.php
@@ -27,6 +27,5 @@ class ForgotPasswordController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('guest');
     }
 }