+++ resources/views/auth/passwords/reset.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app')
+@extends('layouts.laravel')
 
 @section('content')
 <div class="container">