+++ resources/views/layouts/laravel.blade.php
@@ -12,8 +12,8 @@
     <title>{{ $p['prefs.project-name'] }}</title>
 
     <!-- Styles -->
-    <link href="css/vue-multiselect.min.css" rel="stylesheet">
-    <link href="css/app.css" rel="stylesheet">
+    <link href="/css/vue-multiselect.min.css" rel="stylesheet">
+    <link href="/css/app.css" rel="stylesheet">
 </head>
 <body>
     <div id="app" class="d-flex flex-column">